+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Snyk Security Changelog
 
 ## [2.23.0]
+- Support early access of PAT authentication
 - Removed Snyk Code Quality feature. All related settings, configurations, and UI elements have been removed.
 - Removed scanning for vulnerabilities in JavaScript libraries referenced via CDN in HTML files.
 - Removed inline display of vulnerability counts for imports/requires in JavaScript/TypeScript files.